refactor(options): move the font settings into a module of their own
`appearance.tsx` had grown to 952 lines, 394 of them the fonts card and the
picker behind it. They sat in two contiguous runs that the rest of the page
reached through a single `<Fonts />` and that reached back for nothing, so the
seam was already there to cut along.

The code moves verbatim: the only edit is the `export default` the page now
imports it by. `appearance.tsx` drops to 533 lines and sheds 24 imports — the
whole commons line, `clsx`, `Fragment`, `createPortal`, `useRef`, `Modal`,
`NoItems`, `Slider` and the font services among them.

The stylesheet splits along the same line, the `@container` block included: it
holds nothing but `.font-option` rules, so it belongs beside them.

The page's own spec stands the card in for rather than repeating the mocks it
needs, which is the shape of the `OptionsPageHeader` stub already beside it. That
lets it drop the custom-fonts and font-service mocks along with the `document.fonts`
stub, and leaves the fonts spec free to render the card on its own.
