+++ apps/client/src/widgets/type_widgets/options/appearance_fonts.css
@@ -1,0 +1,162 @@
+/* Font Option Row */
+.font-option {
+    cursor: pointer;
+
+    /* Shown rather than hidden with custom fonts off, so the settings can still be read. */
+    &.disabled {
+        opacity: 0.5;
+        pointer-events: none;
+    }
+}
+
+.font-option-preview {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    color: var(--muted-text-color);
+}
+
+/* The specimen is drawn at the user's own family and size, so it can outgrow the screen: it takes
+   the line below the name earlier than an ordinary option does. Measured against the card, which
+   `Card.css` makes a container. */
+@container (max-width: 600px) {
+    .font-option {
+        flex-direction: column;
+        align-items: stretch;
+    }
+
+    .font-option .font-option-preview {
+        justify-content: space-between;
+    }
+
+    .font-option .font-option-specimen {
+        /* `anywhere`: a family name is often one word, with no break the browser would take. */
+        min-width: 0;
+        overflow-wrap: anywhere;
+    }
+}
+
+/* Font Picker Modal */
+body.desktop .modal.font-picker-modal {
+    .modal-content-with-sidebar {
+        height: 60vh;
+    }
+
+    /* On the sidebar rather than on the list, which a search matching nothing takes away. */
+    .modal-sidebar {
+        width: 280px;
+    }
+}
+
+.modal.font-picker-modal {
+    /* Keep the sidebar background within the rounded corners. */
+    .modal-content {
+        overflow: hidden;
+    }
+
+    .modal-body {
+        padding: 0;
+        display: flex;
+        flex-direction: column;
+    }
+
+    .modal-sidebar {
+        background-color: var(--card-background-color);
+    }
+
+    .modal-sidebar .dropdown-menu.static {
+        background-color: transparent !important;
+        border: none;
+        border-radius: 0 !important;
+    }
+
+    /* Set here rather than by the field, which takes the room it is given (see SettingsSearch.css).
+       Room below as well as above: the focus ring reaches past the border, and the list beneath it
+       paints over whatever crosses into it. */
+    .settings-search {
+        margin: 8px;
+    }
+
+    .font-picker-list {
+        flex: 1 1 0;
+        min-height: 0;
+        overflow: auto;
+
+        .dropdown-item {
+            font-size: 14px;
+        }
+    }
+
+    .font-picker-empty {
+        flex: 1 1 0;
+        min-height: 0;
+        padding: 16px;
+    }
+
+    .font-picker-settings {
+        flex: 1;
+        display: flex;
+        flex-direction: column;
+        padding: 16px;
+        gap: 16px;
+    }
+}
+
+.font-size-control {
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+
+    label {
+        font-weight: 500;
+        color: var(--muted-text-color);
+        font-size: 12px;
+        text-transform: uppercase;
+    }
+
+    .font-size-description {
+        color: var(--muted-text-color);
+        font-size: 12px;
+    }
+}
+
+.font-size-slider {
+    display: flex;
+    align-items: center;
+    gap: 12px;
+
+    .slider {
+        flex: 1;
+    }
+
+    .font-size-value {
+        min-width: 50px;
+        text-align: right;
+        font-variant-numeric: tabular-nums;
+    }
+}
+
+.font-preview {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+    min-height: 0;
+
+    label {
+        font-weight: 500;
+        color: var(--muted-text-color);
+        font-size: 12px;
+        text-transform: uppercase;
+    }
+}
+
+.font-preview-text {
+    flex: 1;
+    padding: 16px;
+    border: 1px solid var(--main-border-color);
+    border-radius: 6px;
+    background: var(--main-background-color);
+    overflow: auto;
+    line-height: 1.6;
+}
